requirements.txt
Django
djangorestframework
Pillow
python-decouple
drf_yasg
